Add support for completing words instead of lines
I tried to make the changes as less intrusive as possible. This was
achieved by only passing the current word to the completion callback and
expanding all returned completions afterwards. This obviously isn't the
most effective way to implement this (from a memory usage perspective)
but it required only few changes to existing functions.